+++ CHANGES.rst
@@ -6,3 +6,5 @@ Version 0.1
 ===========
 - create installable project using pyscaffold
 - support for python 2.7
+- switch from print to logging
+- use tox for testing python 2 and 3